+++ lib/seforim_library_updater.dart
@@ -14,6 +14,8 @@ export 'src/models/library_update_plan.dart'
     show LibraryUpdatePlan, LibraryUpdatePlanKind, PatchEdge;
 export 'src/models/patch_table_spec.dart'
     show PatchTableSpec, kPatchTablesInFkOrder, kHashTableOrder;
+export 'src/services/content_hash_stamp_store.dart'
+    show ContentHashStamp, ContentHashStampStore;
 export 'src/services/github_library_release_client.dart'
     show GithubLibraryReleaseClient;
 export 'src/services/library_db_recovery_service.dart'